chore: 로또 당첨금액 타입변경
- 로또 1등의 경우 int 범위를 초과하는 금액이 들어가기 때문에 당첨금액 타입을 변경해주었습니다.

+++ src/main/java/lotto/domain/Rank.java
@@ -10,7 +10,7 @@ public enum Rank {
 	NOT(0, 0);
 
 	private final int num;
-	private final int amount;
+	private final long amount;
 
 	Rank(int num, int amount) {
 		this.num = num;
@@ -32,11 +32,11 @@ public enum Rank {
 		return num;
 	}
 
-	public int getAmount() {
+	public long getAmount() {
 		return amount;
 	}
 
-	public Integer calcTotalAmount(Integer winners) {
+	public long calcTotalAmount(Integer winners) {
 		return this.amount * winners;
 	}
 }
